+++ include/seqan3/io/sequence_file/output.hpp
@@ -21,6 +21,7 @@
 #include <range/v3/algorithm/equal.hpp>
 #include <range/v3/view/zip.hpp>
 
+#include <seqan3/core/algorithm/parameter_pack.hpp>
 #include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
 #include <seqan3/core/concept/tuple.hpp>
@@ -730,7 +731,25 @@ protected:
  * \{
  */
 
-//!\brief Deduction of the selected fields, the file format and the stream type.
+//!\brief Deduction guide for given stream and file format.
+template <OStream2                 stream_t,
+          SequenceFileOutputFormat file_format>
+sequence_file_output(stream_t &,
+                     file_format const &)
+    -> sequence_file_output<typename sequence_file_output<>::selected_field_ids,  // default field ids
+                            type_list<file_format>,
+                            typename std::remove_reference_t<stream_t>::char_type>;
+
+//!\overload
+template <OStream2                 stream_t,
+          SequenceFileOutputFormat file_format>
+sequence_file_output(stream_t &&,
+                     file_format const &)
+    -> sequence_file_output<typename sequence_file_output<>::selected_field_ids,  // default field ids.
+                            type_list<file_format>,
+                            typename std::remove_reference_t<stream_t>::char_type>;
+
+//!\brief Deduction guide for given stream, file format and field ids.
 template <OStream2                 stream_t,
           SequenceFileOutputFormat file_format,
           detail::Fields           selected_field_ids>